bowlingball.com
Refund Policy

When you make a purchase on Poshmark, we don't release payment to the seller until you tell us you've received your order as described. You have 3 days after delivery to inform us if the item has been misrepresented by reporting the problem in the Poshmark app or website with supporting photos. If we verify your claim, we'll send you a label to return the order to the seller and refund your payment. All returns must be shipped back within 5 days of approval to be eligible for a refund.

What is covered

If any of the following pertain to your order, please report the problem in the Poshmark app or website immediately:

Undisclosed damage
Incorrect or missing item
Item not as described
Item is not authentic
If no claim is made within 3 days of delivery, payment will be automatically released to the seller. Once payment has been released, all sales are final and no refunds will be provided.

What is not covered

Item does not fit / changed mind
If the item is just not your style or does not fit you, we unfortunately cannot accept a return. You can always re-list the item on Poshmark!

Trades and Offline Transactions
Posh Protect does not cover trades or transactions completed off of the Poshmark platform. When payment for the full value of items is not exchanged through the Poshmark platform, we cannot guarantee that both parties will ship and transact as promised. You bear all risks associated with any such transactions.